chore: pet 정보 반환 매퍼 수정

+++ src/main/java/com/kau/capstone/domain/pet/util/PetMapper.java
@@ -31,7 +31,7 @@ public class PetMapper {
             .petType(pet.getPetType())
             .gender(pet.getGender())
             .name(pet.getName())
-            .weight((float) pet.getWeight())
+            .weight(pet.getWeight())
             .imageUrl(pet.getImageUrl())
             .build();
     }
